refactor: simplified the main function
The decode function take a line of string and tries to parse it into
a valid terminal reponse. It uses two helper functions.

fetchRawData, this function return only the payload of the response and
discard the date and request on the line.

convertHexToASCII take the hex value of the response and convert them
into a string.

We then split the string into each field using character 0x1C as the
delimiter.